requirements.txt
azure-identity
azure-search-documents
azure-storage-blob
bs4
langchain
msal>=0.6.1,<2
openai
pandas
pypdf
PyPDF2
python-docx
python-dotenv
requests>=2,<3
tiktoken
azure-ai-formrecognizer
